feat: move all the non-wasm code into its own little world
