+++ tomb-crypt/src/key_seal/standard/internal.rs
@@ -10,8 +10,9 @@ use openssl::nid::Nid;
 use openssl::pkey::{PKey, Private, Public};
 
 use crate::key_seal::{
-    pretty_fingerprint, AES_KEY_SIZE, ECDH_SECRET_BYTE_SIZE, FINGERPRINT_SIZE, SALT_SIZE, KeySealError,
+    pretty_fingerprint, KeySealError,
 };
+use crate::key_seal::common::{AES_KEY_SIZE, ECDH_SECRET_BYTE_SIZE, FINGERPRINT_SIZE, SALT_SIZE};
 
 pub(crate) fn base64_decode(data: &str) -> Result<Vec<u8>, KeySealError> {
     B64.decode(data).map_err(KeySealError::bad_base64)
